fix: resolve all 27 compilation errors across 13 files
Batch 1 - Remove ghost fields (isDeleted/updatedAt not in QuickEntryEntity):
- SaveQuickEntryUseCase: remove isDeleted, updatedAt params
- DeleteQuickEntryUseCase: convert soft-delete to hard delete
- PayDebtUseCase: remove isDeleted/updatedAt on QuickEntryEntity, add categoryId to DebtPaymentEntity
- GetDashboardDataUseCase: remove isDeleted filter

Batch 2 - Fix Float→Double type mismatches:
- SaveCapturedOrderUseCase: parseConfidence Float→Double
- SaveHistoryDetailUseCase: parseConfidence Float→Double
- CreateDataReviewUseCase: confidence Float→Double, constants 0.85f→0.85
- MatchOrdersUseCase: confidence literals f→Double

Batch 3 - Fix WorkScheduleRepository DAO method names:
- getBetween→getByDateRange, upsert→insert, getOffDayCountBetween→countWorkingDays

Batch 4 - Fix MatchF001F002UseCase:
- getDetailsByHistoryTripIdOnce→getDetailsByHistoryTripId().first()

Batch 5 - Fix SaveScreenSnapshotUseCase:
- imagePath String?→String with fallback

Supporting - Add missing DAO/Repository methods:
- QuickEntryDao: add @Update, @Delete, getSumByDateAndType, getCountByDateAndType
- QuickEntryRepository: add deleteEntry method

+++ app/src/main/java/com/driverfinance/domain/usecase/capture/SaveCapturedOrderUseCase.kt
@@ -27,7 +27,7 @@ class SaveCapturedOrderUseCase @Inject constructor(
         paymentMethod: String?,
         paymentAmount: Int?,
         rawText: String,
-        parseConfidence: Float,
+        parseConfidence: Double,
         sourceSnapshotId: String
     ): SaveResult {
         val now = OffsetDateTime.now().format(DateTimeFormatter.ISO_OFFSET_DATE_TIME)
